admin updates + refactoring
